Move chartbeat to src/app/components & update references
