+++ src/app/components/ChartbeatAnalytics/amp/index.jsx
