+++ src/app/components/ChartbeatAnalytics/amp/index.test.jsx
