+++ src/app/components/ChartbeatAnalytics/canonical/index.jsx
