+++ src/app/components/ChartbeatAnalytics/canonical/index.test.jsx
