+++ src/app/components/ChartbeatAnalytics/index.jsx
@@ -3,7 +3,7 @@ import useToggle from '#hooks/useToggle';
 import { UserContext } from '#contexts/UserContext';
 import { RequestContext } from '#contexts/RequestContext';
 import { pageDataPropType } from '#models/propTypes/data';
-import { ServiceContext } from '../../../contexts/ServiceContext';
+import { ServiceContext } from '../../contexts/ServiceContext';
 import AmpChartbeatBeacon from './amp';
 import { getConfig } from './utils';
