+++ src/app/components/ChartbeatAnalytics/index.test.jsx
@@ -2,14 +2,14 @@ import React from 'react';
 import { node, string, shape } from 'prop-types';
 import { render } from '@testing-library/react';
 import { ARTICLE_PAGE } from '#app/routes/utils/pageTypes';
-import { RequestContextProvider } from '../../../contexts/RequestContext';
-import { ServiceContextProvider } from '../../../contexts/ServiceContext';
-import { ToggleContext } from '../../../contexts/ToggleContext';
-import { UserContext } from '../../../contexts/UserContext';
+import { RequestContextProvider } from '../../contexts/RequestContext';
+import { ServiceContextProvider } from '../../contexts/ServiceContext';
+import { ToggleContext } from '../../contexts/ToggleContext';
+import { UserContext } from '../../contexts/UserContext';
 import ChartbeatAnalytics from '.';
 import * as testUtils from './utils';
 import * as amp from './amp';
-import { localBaseUrl } from '../../../../testHelpers/config';
+import { localBaseUrl } from '../../../testHelpers/config';
 import frontPageData from '../../../../../data/news/frontpage';
 
 const defaultToggleState = {
