+++ src/app/components/ChartbeatAnalytics/utils/index.js
