+++ src/app/components/ChartbeatAnalytics/utils/index.test.js
