+++ src/app/pages/IdxPage/IdxPage.jsx
@@ -23,8 +23,8 @@ import RadioScheduleContainer from '#containers/RadioSchedule';
 import IndexPageSection from '#containers/IndexPageSection';
 import idxPageDataPropTypes from '#models/propTypes/idxPage';
 import ATIAnalytics from '#containers/ATIAnalytics';
-import ChartbeatAnalytics from '#containers/ChartbeatAnalytics';
 import ComscoreAnalytics from '#containers/ComscoreAnalytics';
+import ChartbeatAnalytics from '../../components/ChartbeatAnalytics';
 import { ServiceContext } from '../../contexts/ServiceContext';
 import MetadataContainer from '../../components/Metadata';
 import { GHOST } from '../../components/ThemeProvider/palette';
